requirements.txt
# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Machine Learning & Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# LLM Integration
anthropic>=0.18.0

# Configuration Management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
nflreadpy>=0.1.0
pandas>=2.0.0

pyarrow

beautifulsoup4>=4.12.0
requests>=2.31.0

# CLI and formatting
rich>=13.0.0
